requirements.txt
flask>=2.3.0
opencv-python>=4.8.0
numpy>=1.24.0
boto3>=1.28.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
Pillow>=9.0.0
